unsigned->uint64_t for AbsCommonStrideInBytes

+++ llvm/include/llvm/Analysis/LoopAccessAnalysis.h
@@ -494,11 +494,11 @@ struct PointerDiffInfo {
   const SCEV *SrcStart;
   const SCEV *SinkStart;
   unsigned AccessSize;
-  unsigned AbsCommonStrideInBytes;
+  uint64_t AbsCommonStrideInBytes;
   bool NeedsFreeze;
 
   PointerDiffInfo(const SCEV *SrcStart, const SCEV *SinkStart,
-                  unsigned AccessSize, unsigned AbsCommonStrideInBytes,
+                  unsigned AccessSize, uint64_t AbsCommonStrideInBytes,
                   bool NeedsFreeze)
       : SrcStart(SrcStart), SinkStart(SinkStart), AccessSize(AccessSize),
         AbsCommonStrideInBytes(AbsCommonStrideInBytes),
